Bugfixes, 1.5.2 Release
- .syntactic_sweets.properties.decorators:
    - made into its own sub-package
       - added `paths` module
           - moved `validate_path` to `paths` module
           - added `provision_path` to `paths` module
       - added `type_validation` module
           - moved `validate_type` to the `type_validation` module

- version bump 1.5.1 -> 1.5.2
